+++ ddsc_socket/__init__.py
@@ -1,2 +1,2 @@
-# (c) Fugro GeoServices. MIT licensed, see LICENSE.rst.
-# package
+# (c) Fugro GeoServices, Nelen & Schuurmans. MIT licensed, see LICENSE.rst.
+# -*- coding: utf-8 -*-