+++ CHANGES.rst
@@ -6,3 +6,7 @@ Changelog of ddsc-socket
 ----------------
 
 - Initial project structure created with nensskel 1.30.
+
+- Upgrade bootstrap.py.
+
+- Prepare for lizard-nxt.